requirements.txt
sentence-transformers>=2.2.0
hdbscan>=0.8.33
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
datasets>=2.14.0
tqdm>=4.65.0
openai>=1.0.0
python-dotenv>=1.0.0
jupyter>=1.0.0
ipykernel>=6.25.0
